requirements.txt
Flask==2.3.3
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
mysql-connector-python==8.1.0